requirements.txt
numpy==2.2.0
aiohttp==3.11.10
python-dotenv==1.0.1
aiogram==3.17.0
asyncio==3.4.3
typing==3.7.4.3